add tests to cars service

+++ src/main/java/com/course/cars/domain/CarsService.java
@@ -4,9 +4,7 @@ import com.course.cars.domain.dto.CarDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
-import org.springframework.web.bind.annotation.RequestBody;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,24 +26,30 @@ public class CarsService {
         return repository.findAllByType(type).stream().map(CarDTO::create).toList();
     }
 
-    public Car create(Car car) {
-        return repository.save(car);
+    public CarDTO create(Car car) {
+        Assert.isNull(car.getId(), "Impossible to insert data.");
+
+        return CarDTO.create(repository.save(car));
     }
 
-    public Car update(Long id, Car car) {
+    public CarDTO update(Long id, Car car) {
         Assert.notNull(id, "Impossible to update data.");
 
-        return repository.findById(id).map(c -> {
+        return CarDTO.create(repository.findById(id).map(c -> {
             c.setName(car.getName());
             c.setType(car.getType());
 
             repository.save(c);
             return c;
-        }).orElseThrow(() -> new RuntimeException("Impossible to update data."));
+        }).orElseThrow(() -> new RuntimeException("Impossible to update data.")));
 
     }
 
-    public void delete(Long id) {
-        repository.deleteById(id);
+    public boolean delete(Long id) {
+        if (getById(id).isPresent()) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
